+++ src/rsXML/mindstorms.cpp
@@ -4,7 +4,7 @@
 using namespace rsXML;
 using namespace rsMindstorms;
 
-Mindstorms::Mindstorms(int form, bool trace) : rsRobots::Robot(form), rsRobots::Mindstorms(form), rsXML::Robot(trace) {
+Mindstorms::Mindstorms(short form, bool trace) : rsRobots::Robot(form), rsRobots::Mindstorms(form), rsXML::Robot(trace) {
 	_a.allocate(Bodies::Num_Joints);
 	for (unsigned int i = 0; i < _a.size(); i++) {
 		_a[i] = 0;
